rename street_name to street_name_part
