+++ src/locales/uz_UZ_latin/location/street_name_part.ts
